Convert FitImage into TSX file
* I tried to limit functional changes to minimum -- focused on adding typings
